requirements.txt
groundx>=1.3.22
crawl4ai>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.0.0
lxml>=5.0.0
nest-asyncio>=1.6.0
